fix(extension): stop the panel wrapper fighting the popup's own sizing
The spike passed — --font-sans and --accent resolved and Tailwind utilities
applied inside the shadow root, so the :host mirror works and Twitch's
CSP/trusted-types does not block the style injection. Architecture settled.

But the panel rendered badly, and the wrapper caused it. <Popup>'s root is
already h-[600px] w-[400px] flex flex-col overflow-hidden and owns its internal
scrolling, so the container's width/max-height/overflow were redundant and
actively harmful: the diagnostic strip pushed total height past 600px, which
raised an outer vertical scrollbar, which ate ~15px of the 400px the popup
insists on, which overflowed horizontally and clipped the popup's own text.

The container now carries position and rounding only. Removes the StyleProbe
strip, and renames the module from inPagePanelSpike.tsx to inPagePanel.tsx
since it is no longer a throwaway harness.
